moving and shuffling
