+++ Sensors/DHT22_fdrs/DHT22_fdrs.ino
@@ -3,8 +3,8 @@
 
 // Modified by Timm Bogner for Farm Data Relay System -- Untested because I don't have a DHT sensor onhand. 
 
-
-#include "fdrs_sensor.h"
+#include "sensor_setup.h"
+#include <fdrs_sensor.h>
 #include "DHT.h"
 
 #define DHTPIN 2     // Digital pin connected to the DHT sensor
