+++ Sensors/DS18B20_fdrs/DS18B20_fdrs.ino
@@ -5,7 +5,8 @@
 
 #define ONE_WIRE_BUS  13   //Pin that the DS18B20 is connected to
 
-#include "fdrs_sensor.h"
+#include "sensor_setup.h"
+#include <fdrs_sensor.h>
 #include <OneWire.h>
 #include <DallasTemperature.h>
