+++ Sensors/LilyGo_HiGrow_32/LilyGo_HiGrow_32.ino
@@ -14,7 +14,8 @@
 #define USER_BUTTON         35
 #define DS18B20_PIN         21
 
-#include "fdrs_sensor.h"
+#include "sensor_setup.h"
+#include <fdrs_sensor.h>
 #include <BH1750.h>
 #include <Adafruit_BME280.h>
