+++ fdrs_globals.h
